added permanent key mapping configuration
